# services/investment-service/docs/openapi.yaml
openapi: 3.1.0
info:
  title: Investment Service
  description: Investment service API for wealth manager platform. Manages mutual
    fund investments and portfolio analytics.
  contact:
    name: Wealth Manager Team
    url: https://github.com/sri-akshat/wealth-manager
  license:
    name: Private
  version: 1.0.0
paths:
  /:
    get:
      tags:
      - system
      summary: Root
      description: 'Root endpoint that provides basic service information.


        Returns service name, version, and status.'
      operationId: root__get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
  /portfolio/summary:
    get:
      tags:
      - portfolio
      summary: Get Portfolio Summary
      description: 'Get summary of user''s investment portfolio.


        Returns total investment, current value, returns, and asset allocation.'
      operationId: get_portfolio_summary_portfolio_summary_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PortfolioSummary'
      security:
      - OAuth2PasswordBearer: []
  /initialize-sample-data:
    post:
      tags:
      - system
      summary: Initialize Data
      description: 'Initialize sample mutual fund data for testing.


        This endpoint adds a set of predefined mutual funds to the database

        if they don''t already exist.'
      operationId: initialize_data_initialize_sample_data_post
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
  /investments/:
    post:
      tags:
      - investments
      summary: Create Investment
      description: 'Create a new investment for the user.


        Takes investment details and creates a new investment record.'
      operationId: create_investment_investments__post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InvestmentCreate'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InvestmentResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - OAuth2PasswordBearer: []
  /portfolio/investments:
    get:
      tags:
      - portfolio
      summary: Get Portfolio Investments
      description: 'Get detailed list of user''s investments.


        Returns a list of all investments in the user''s portfolio with calculated

        returns and current values.'
      operationId: get_portfolio_investments_portfolio_investments_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PortfolioInvestmentList'
      security:
      - OAuth2PasswordBearer: []
  /portfolio/analytics:
    get:
      tags:
      - portfolio
      summary: Get Portfolio Analytics
      description: 'Get detailed portfolio analytics.


        Returns portfolio summary and detailed investment information.'
      operationId: get_portfolio_analytics_portfolio_analytics_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PortfolioAnalytics'
      security:
      - OAuth2PasswordBearer: []
  /investments/update-navs:
    post:
      tags:
      - investments
      summary: Update Investment Navs
      description: 'Update current NAVs and values for all investments.


        This endpoint recalculates the current value of all investments

        based on the latest NAV data. In a production environment,

        this would fetch real-time NAV data from an external API.'
      operationId: update_investment_navs_investments_update_navs_post
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
      security:
      - OAuth2PasswordBearer: []
  /health:
    get:
      tags:
      - system
      summary: Health Check
      description: 'Check if the service is healthy.


        Returns a simple status message indicating the service is operational.'
      operationId: health_check_health_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
components:
  schemas:
    FundCategory:
      type: string
      enum:
      - equity
      - debt
      - hybrid
      - liquid
      - index
      title: FundCategory
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
    InvestmentCreate:
      properties:
        fund_id:
          type: integer
          title: Fund Id
        purchase_amount:
          type: number
          exclusiveMinimum: 0.0
          title: Purchase Amount
      type: object
      required:
      - fund_id
      - purchase_amount
      title: InvestmentCreate
    InvestmentResponse:
      properties:
        id:
          type: integer
          title: Id
        user_id:
          type: integer
          title: User Id
        fund_id:
          type: integer
          title: Fund Id
        units:
          type: number
          title: Units
        purchase_nav:
          type: number
          title: Purchase Nav
        current_nav:
          type: number
          title: Current Nav
        purchase_amount:
          type: number
          title: Purchase Amount
        current_value:
          type: number
          title: Current Value
        purchase_date:
          type: string
          format: date-time
          title: Purchase Date
        status:
          $ref: '#/components/schemas/InvestmentStatus'
      type: object
      required:
      - id
      - user_id
      - fund_id
      - units
      - purchase_nav
      - current_nav
      - purchase_amount
      - current_value
      - purchase_date
      - status
      title: InvestmentResponse
    InvestmentStatus:
      type: string
      enum:
      - pending
      - completed
      - failed
      - cancelled
      title: InvestmentStatus
    MessageResponse:
      properties:
        message:
          type: string
          title: Message
        service:
          type: string
          title: Service
        version:
          type: string
          title: Version
        status:
          type: string
          title: Status
      type: object
      required:
      - message
      - service
      - version
      - status
      title: MessageResponse
      description: A simple message response schema.
    PortfolioAnalytics:
      properties:
        summary:
          $ref: '#/components/schemas/PortfolioSummary'
        investments:
          items:
            $ref: '#/components/schemas/PortfolioInvestment'
          type: array
          title: Investments
      type: object
      required:
      - summary
      - investments
      title: PortfolioAnalytics
    PortfolioInvestment:
      properties:
        id:
          type: integer
          title: Id
        fund_name:
          type: string
          title: Fund Name
        category:
          $ref: '#/components/schemas/FundCategory'
        units:
          type: number
          title: Units
        purchase_nav:
          type: number
          title: Purchase Nav
        current_nav:
          type: number
          title: Current Nav
        purchase_amount:
          type: number
          title: Purchase Amount
        current_value:
          type: number
          title: Current Value
        returns:
          type: number
          title: Returns
        returns_percentage:
          type: number
          title: Returns Percentage
        purchase_date:
          type: string
          format: date-time
          title: Purchase Date
      type: object
      required:
      - id
      - fund_name
      - category
      - units
      - purchase_nav
      - current_nav
      - purchase_amount
      - current_value
      - returns
      - returns_percentage
      - purchase_date
      title: PortfolioInvestment
    PortfolioInvestmentList:
      properties:
        investments:
          items:
            $ref: '#/components/schemas/PortfolioInvestment'
          type: array
          title: Investments
      type: object
      required:
      - investments
      title: PortfolioInvestmentList
      description: A list of portfolio investments.
    PortfolioSummary:
      properties:
        total_investment:
          type: number
          title: Total Investment
        current_value:
          type: number
          title: Current Value
        total_returns:
          type: number
          title: Total Returns
        returns_percentage:
          type: number
          title: Returns Percentage
        number_of_investments:
          type: integer
          title: Number Of Investments
        asset_allocation:
          additionalProperties:
            type: number
          type: object
          title: Asset Allocation
      type: object
      required:
      - total_investment
      - current_value
      - total_returns
      - returns_percentage
      - number_of_investments
      - asset_allocation
      title: PortfolioSummary
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
  securitySchemes:
    OAuth2PasswordBearer:
      type: oauth2
      flows:
        password:
          scopes: {}
          tokenUrl: token
tags:
- name: portfolio
  description: Operations related to investment portfolios
- name: investments
  description: Operations for managing individual investments
- name: system
  description: System maintenance operations